chore(*): prepare for Banana ML serverless - 002
After taking a long detour to understand how the serverless framework
works on Banana ML, this project was updated accordingly. The biggest
difference is the fact that it's not pulling from my own, custom image,
as opposed to from Pytorch's v1.11.0 (CUDA v11.3). My image comes pre-
installed with a bunch of requirements and should include at least a few
of the models that are currently hosted on HuggingFace.

Potential caveats:
- Not all models have been downloaded
- Some dependencies may have been skipped on accident
- The actual App and API remain unvetted at this stage
  (Those are to be tested locally, potentially in parallel with Banana)
- There's no access to the outputs; not even a way to list them (yet)

Let's rock! :guitar: